flask_app/requirements.txt
numpy
flask
celery
requests
opencv-python==4.4.0.44
pillow
redis
future
imgaug
shapely
scipy
validators